+++ src/jtree.js
@@ -13,6 +13,6 @@ jtree.TreeNode = TreeNode
 jtree.NonTerminalNode = GrammarBackedNonTerminalNode
 jtree.TerminalNode = GrammarBackedTerminalNode
 
-jtree.getVersion = () => "14.4.0"
+jtree.getVersion = () => "14.5.0"
 
 module.exports = jtree